Update MenuScreen
+ Fix Rating
+ Optimize MenuScreen
+ MenuDetail script
+ Class DishContent
