+++ src/ARMenu/Assets/MenuAssets/DishContent.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0da86db791ca5d14d912878fa29ded45
-timeCreated: 1513051000
+guid: 8645573a9c4c4f146b3e734fddf494e0
+timeCreated: 1513203775
 licenseType: Free
 MonoImporter:
   externalObjects: {}
