+++ src/ARMenu/Assets/MenuAssets/MenuDetail.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: 9f88d374b49ebca40a999f0e86040389
-timeCreated: 1513073402
+timeCreated: 1513206899
 licenseType: Free
 NativeFormatImporter:
   externalObjects: {}
