+++ src/ARMenu/Assets/MenuAssets/MenuDetailControl.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e0859060238c34b46851acf9c01ee3b7
+timeCreated: 1513202731
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
